+++ projects/output-and-error-handling/server/main.go
@@ -8,6 +8,7 @@ import (
 	"strconv"
 	"time"
 )
+const ServerURL = "http://127.0.0.1:8080"
 
 func main() {
 	http.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
